refactor ModelVersionHeader
