+++ webui/react/src/components/UseNotebookModalComponent.module.scss
@@ -1,0 +1,4 @@
+.topLine {
+  display: flex;
+  justify-content: space-between;
+}
